+++ scripts/check-changeset/src/index.ts
@@ -13,9 +13,6 @@ async function run() {
   );
   for (const changeset of changesets) {
     const { id, releases, summary } = changeset;
-    if (id === 'rotten-carrots-joke') {
-      continue;
-    }
     releases.forEach(release => {
       if (release.type === 'major') {
         throw Error(`packages ${release.name} not allow bump major version`);