+++ packages/cli/plugin-bff/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @modern-js/plugin-bff
 
+## 2.0.2
+
+### Patch Changes
+
+- @modern-js/utils@2.0.2
+- @modern-js/bff-core@2.0.2
+- @modern-js/create-request@2.0.2
+- @modern-js/server-utils@2.0.2
+- @modern-js/babel-compiler@2.0.2
+
+## 2.0.1
+
+### Patch Changes
+
+- @modern-js/bff-core@2.0.1
+- @modern-js/create-request@2.0.1
+- @modern-js/server-utils@2.0.1
+- @modern-js/babel-compiler@2.0.1
+- @modern-js/utils@2.0.1
+
+## 2.0.0
+
+### Major Changes
+
+- dda38c9c3e: chore: v2
+
+### Patch Changes
+
+- Updated dependencies [edd1cfb1af]
+- Updated dependencies [dda38c9c3e]
+- Updated dependencies [ffb2ed4]
+- Updated dependencies [bbe4c4ab64]
+  - @modern-js/utils@2.0.0
+  - @modern-js/bff-core@2.0.0
+  - @modern-js/create-request@2.0.0
+  - @modern-js/server-utils@2.0.0
+  - @modern-js/babel-compiler@2.0.0
+
 ## 2.0.0-beta.7
 
 ### Major Changes