+++ packages/generator/generator-cases/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @modern-js/generator-cases
 
+## 3.0.2
+
+### Patch Changes
+
+- Updated dependencies [4568607]
+  - @modern-js/generator-common@3.0.2
+
+## 3.0.1
+
+### Patch Changes
+
+- @modern-js/generator-common@3.0.1
+
+## 3.0.0
+
+### Major Changes
+
+- dda38c9c3e: chore: v2
+
+### Patch Changes
+
+- Updated dependencies [dda38c9c3e]
+  - @modern-js/generator-common@3.0.0
+
 ## 3.0.0-beta.6
 
 ### Major Changes