+++ packages/generator/generators/base-generator/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @modern-js/base-generator
 
+## 3.0.2
+
+## 3.0.1
+
+## 3.0.0
+
+### Major Changes
+
+- dda38c9c3e: chore: v2
+
+### Patch Changes
+
+- f680410886: feat: upgrade ESLint to 8.x version
+
+  feat: 升级 ESLint 到 8.x 版本
+
 ## 3.0.0-beta.6
 
 ### Major Changes