+++ packages/runtime/plugin-testing/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @modern-js/plugin-testing
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [39988b2]
+  - @modern-js/runtime@2.0.2
+  - @modern-js/prod-server@2.0.2
+  - @modern-js/utils@2.0.2
+  - @modern-js/babel-preset-app@2.0.2
+  - @modern-js/babel-compiler@2.0.2
+  - @modern-js/plugin@2.0.2
+
+## 2.0.1
+
+### Patch Changes
+
+- @modern-js/babel-preset-app@2.0.1
+- @modern-js/runtime@2.0.1
+- @modern-js/prod-server@2.0.1
+- @modern-js/babel-compiler@2.0.1
+- @modern-js/plugin@2.0.1
+- @modern-js/utils@2.0.1
+
+## 2.0.0
+
+### Major Changes
+
+- dda38c9c3e: chore: v2
+
+### Patch Changes
+
+- Updated dependencies [c9e800d39a]
+- Updated dependencies [edd1cfb1af]
+- Updated dependencies [6bda14ed71]
+- Updated dependencies [dda38c9c3e]
+- Updated dependencies [8b8e1bb571]
+- Updated dependencies [ffb2ed4]
+- Updated dependencies [bbe4c4ab64]
+  - @modern-js/runtime@2.0.0
+  - @modern-js/prod-server@2.0.0
+  - @modern-js/utils@2.0.0
+  - @modern-js/babel-preset-app@2.0.0
+  - @modern-js/babel-compiler@2.0.0
+  - @modern-js/plugin@2.0.0
+
 ## 2.0.0-beta.7
 
 ### Major Changes